requirements.txt
jax==0.4.26
antlr4-python3-runtime==4.9.3 ; python_version >= "3.10" and python_version < "4.0"
certifi==2024.7.4 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0"
contourpy==1.2.1 ; python_version >= "3.10" and python_version < "4.0"
cycler==0.12.1 ; python_version >= "3.10" and python_version < "4.0"
dill==0.3.8 ; python_version >= "3.10" and python_version < "4.0"
equinox==0.11.4 ; python_version >= "3.10" and python_version < "4.0"
fonttools==4.53.1 ; python_version >= "3.10" and python_version < "4.0"
hydra-core==1.3.2 ; python_version >= "3.10" and python_version < "4.0"
jaxlib==0.4.26 ; python_version >= "3.10" and python_version < "4.0"
jaxonloader==0.9.1 ; python_version >= "3.10" and python_version < "4.0"
jaxtyping==0.2.33 ; python_version >= "3.10" and python_version < "4.0"
joblib==1.4.2 ; python_version >= "3.10" and python_version < "4.0"
kiwisolver==1.4.5 ; python_version >= "3.10" and python_version < "4.0"
lineax==0.0.5 ; python_version >= "3.10" and python_version < "4.0"
matplotlib==3.9.1 ; python_version >= "3.10" and python_version < "4.0"
ml-dtypes==0.4.0 ; python_version >= "3.10" and python_version < "4.0"
numpy==1.26.4 ; python_version >= "3.10" and python_version < "4.0"
omegaconf==2.3.0 ; python_version >= "3.10" and python_version < "4.0"
opt-einsum==3.3.0 ; python_version >= "3.10" and python_version < "4.0"
optimistix==0.0.7 ; python_version >= "3.10" and python_version < "4.0"
packaging==24.1 ; python_version >= "3.10" and python_version < "4.0"
pandas==2.2.2 ; python_version >= "3.10" and python_version < "4.0"
pillow==10.4.0 ; python_version >= "3.10" and python_version < "4.0"
pyparsing==3.1.2 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.10" and python_version < "4.0"
pytz==2024.1 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.10" and python_version < "4.0"
scikit-learn==1.5.1 ; python_version >= "3.10" and python_version < "4.0"
scipy==1.14.0 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
threadpoolctl==3.5.0 ; python_version >= "3.10" and python_version < "4.0"
typeguard==2.13.3 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.10" and python_version < "4.0"
ucimlrepo==0.0.7 ; python_version >= "3.10" and python_version < "4.0"
uqtestfuns==0.4.1 ; python_version >= "3.10" and python_version < "4.0"
